Extract prefab menu items to new class
